Updated Readme with Sprint 1 assignments
Drew and Austin are working on Client-side UI with Android XML.
Andres will be working on formatting transaction objects for display in
the Android UI.
Adam will be working on the Client-side datastore for local transaction
and budget info.
Seth will be working on the server-side database and establishing the
required entities. Also seth can start filling in the vendor list on the
server's database.